Create check.js
The isUidUnique function takes a UID as its argument and returns a boolean indicating whether the UID already exists in the database. You can then call createUid with this function to generate a new UID that is guaranteed to be unique.

In this example, db.checkUidExists is a hypothetical function that checks whether a given UID already exists in the database and returns a boolean. The isUidUnique function simply negates the result of db.checkUidExists, so it returns true if the UID does not exist in the database and false if it does.

The createUid function generates a new UID using the same algorithm as before, but it continues generating UIDs until it finds one that is not already in the database.